feat: unify course sharing for teachers and learners
Changed:
Unified the teacher and learner share menus, refined translated labels, and used default copy-success toasts with learner analytics coverage.

Benefit:
Teachers and learners can share links and copy poster prompts through a consistent interface.
