+++ src/i18n/ar-SA/common/core.json
@@ -36,7 +36,7 @@
   "pageResourcesUpdatedRefreshing": "تم تحديث موارد الصفحة. جارٍ التحديث تلقائيًا...",
   "posterCopy": "نسخ تعليمات الملصق",
   "posterCopyFailed": "تعذر النسخ · انسخ يدويًا",
-  "posterNextStep": "تم النسخ · الصق في مساعد لتوليد الصور",
+  "posterNextStep": "تم النسخ. الصق النص في مساعد ذكاء اصطناعي لإنشاء ملصق",
   "posterPrompt": "أنشئ مباشرة ملصق دورة كاملًا وجاهزًا للمشاركة على وسائل التواصل باستخدام المعلومات التالية.\n\n## معلومات الدورة\n{courseContent}\n\n## المتطلبات\n- اجعل القارئ يفهم سريعًا موضوع الدورة ولمن تناسب ولماذا تستحق الاهتمام، مع إبراز الفوائد التي سيحصل عليها.\n- احتفظ باسم الدورة الدقيق، مع عنوان بارز واحد ونقطتين أو ثلاث نقاط رئيسية. اختر نتائج تعلم أو تجارب أو أعمالًا أو مزايا وفق طبيعة الدورة، دون إضافة نقاط قسرية.\n- استخدم لغة طبيعية وملموسة مع تقليل المصطلحات والفقرات الطويلة. اجعل النص موجزًا ومنظمًا وسهل القراءة على الهاتف.\n- افتراضيًا استخدم تنسيقًا رأسيًا 3:4 وأسلوبًا يناسب الدورة. استخدم تباينًا واضحًا ومساحات كافية وألوان إبراز قليلة ليظل الموضوع معروفًا في الصورة المصغرة.\n- رتّب النص والتصميم بعناية. إذا غابت معلومات ضرورية لاتخاذ قرار، فاسألني.\n- استخدم المعلومات المدعومة فقط. لا تختلق حقائق أو وعودًا بالنتائج.\n- استخدم الرابط الأصلي. أضف رمز QR حقيقيًا فقط إذا أمكن إنشاؤه والتحقّق منه؛ وإلا فاترك مساحة دون رسم رمز وهمي.\n\n## المخرجات\nأنشئ ملصقًا واحدًا مع نص مشاركة قصير والرابط الأصلي للدورة.\nإذا تعذّر توليد الصور، قدّم نص الملصق وتعليمات جاهزة لأداة توليد الصور.",
   "posterPromptLabel": "توجيه الملصق",
   "profileOnboardingManagement": "دليل التخصيص",
@@ -53,7 +53,7 @@
   "share": "مشاركة",
   "shareContentCopied": "تم نسخ محتوى المشاركة",
   "shareCourse": "مشاركة الدورة",
-  "shareCourseLink": "مشاركة رابط الدورة",
+  "shareCourseLink": "مشاركة الرابط",
   "shareCourseMessage": "أنصحك بتعلّم «{courseName}» وتجربة تعلّم شخصي فردي.",
   "shareFailed": "تعذّرت المشاركة، يُرجى المحاولة لاحقًا",
   "shifu": "الدورة",
