+++ src/i18n/en-US/common/core.json
@@ -36,7 +36,7 @@
   "pageResourcesUpdatedRefreshing": "Page resources have been updated. Refreshing automatically...",
   "posterCopy": "Copy poster prompt",
   "posterCopyFailed": "Copy failed · Copy manually",
-  "posterNextStep": "Copied · Paste into an image AI assistant",
+  "posterNextStep": "Copied. Paste into an AI assistant to create a poster",
   "posterPrompt": "Generate a complete course poster ready to share on social media using the information below.\n\n## Course material\n{courseContent}\n\n## Requirements\n- Help readers quickly understand what the course is, who it is for and why it is worth exploring, emphasizing the benefits to them.\n- Keep the exact course name. Create one prominent headline and 2–3 key points, choosing learning outcomes, experiences, projects or distinctive features to suit the course. Use fewer if appropriate.\n- Use natural, concrete language with little jargon and no long paragraphs. Keep the text concise, clearly structured and readable on a phone.\n- Default to a 3:4 portrait layout with a style suited to the course. Use clear contrast, ample whitespace and a few accent colors so the topic remains recognizable in a thumbnail.\n- Arrange the copy and design thoughtfully. If information essential to a decision is missing, ask me.\n- Use only supported information. Do not invent facts or promises of results.\n- Use the original course link. Include a real QR code only if you can generate and verify it; otherwise leave space for one without drawing a fake code.\n\n## Output\nGenerate one poster and provide a short accompanying social caption with the original course link.\nIf you cannot generate images, provide the poster copy and a ready-to-use image generation prompt.",
   "posterPromptLabel": "Poster prompt",
   "profileOnboardingManagement": "Personalization Guide",
@@ -53,7 +53,7 @@
   "share": "Share",
   "shareContentCopied": "Share content copied",
   "shareCourse": "Share course",
-  "shareCourseLink": "Share course link",
+  "shareCourseLink": "Share link",
   "shareCourseMessage": "I recommend learning “{courseName}” for a personalized one-on-one learning experience.",
   "shareFailed": "Unable to share. Please try again later",
   "shifu": "Course",
