+++ src/i18n/fr-FR/common/core.json
@@ -34,9 +34,9 @@
   "orderManagement": "Gestion des commandes",
   "packageManagement": "Gestion des offres",
   "pageResourcesUpdatedRefreshing": "Les ressources de la page ont été mises à jour. Actualisation automatique...",
-  "posterCopy": "Copier la consigne",
+  "posterCopy": "Copier le prompt d’affiche",
   "posterCopyFailed": "Échec de copie · Copiez manuellement",
-  "posterNextStep": "Copié · Collez dans une IA de création d’images",
+  "posterNextStep": "Copié. Collez dans un assistant IA pour créer une affiche",
   "posterPrompt": "Génère directement une affiche de cours complète pour les réseaux sociaux à partir des informations suivantes.\n\n## Informations du cours\n{courseContent}\n\n## Consignes\n- Permets de comprendre rapidement le sujet du cours, son public et son intérêt, en mettant en avant les bénéfices pour le lecteur.\n- Conserve le nom exact du cours. Propose une accroche visible et 2 à 3 points clés : acquis, expériences, réalisations ou particularités selon le cours. N’en ajoute pas artificiellement.\n- Utilise un langage naturel et concret, peu de jargon et aucun long paragraphe. Le texte doit être concis, structuré et lisible sur téléphone.\n- Par défaut, adopte un format portrait 3:4 et un style adapté au cours. Utilise un contraste net, des espaces généreux et quelques couleurs d’accent pour que le thème reste identifiable en miniature.\n- Organise judicieusement les textes et le design. Si une information indispensable à une décision manque, pose-moi la question.\n- Utilise uniquement les informations étayées. N’invente ni faits ni promesses de résultats.\n- Utilise le lien original. Ajoute un vrai QR code uniquement si tu peux le créer et le vérifier ; sinon, réserve un emplacement sans dessiner de faux code.\n\n## Résultat\nGénère une affiche et une courte légende de partage accompagnée du lien original.\nSi tu ne peux pas générer d’image, fournis les textes et une consigne prête à utiliser dans un outil de génération d’images.",
   "posterPromptLabel": "Prompt d’affiche",
   "profileOnboardingManagement": "Guide de personnalisation",
@@ -53,7 +53,7 @@
   "share": "Partager",
   "shareContentCopied": "Contenu à partager copié",
   "shareCourse": "Partager le cours",
-  "shareCourseLink": "Partager le lien du cours",
+  "shareCourseLink": "Partager le lien",
   "shareCourseMessage": "Je vous recommande de suivre « {courseName} » pour découvrir un apprentissage individuel et personnalisé.",
   "shareFailed": "Échec du partage. Veuillez réessayer plus tard",
   "shifu": "Cours",
