+++ src/i18n/th-TH/common/core.json
@@ -36,7 +36,7 @@
   "pageResourcesUpdatedRefreshing": "ทรัพยากรของหน้าได้รับการอัปเดตแล้ว กำลังรีเฟรชโดยอัตโนมัติ...",
   "posterCopy": "คัดลอกพรอมต์โปสเตอร์",
   "posterCopyFailed": "คัดลอกไม่สำเร็จ · คัดลอกด้วยตนเอง",
-  "posterNextStep": "คัดลอกแล้ว · วางใน AI สร้างภาพ",
+  "posterNextStep": "คัดลอกแล้ว วางในผู้ช่วย AI เพื่อสร้างโปสเตอร์",
   "posterPrompt": "สร้างโปสเตอร์หลักสูตรที่สมบูรณ์พร้อมแชร์บนโซเชียลมีเดียโดยตรงจากข้อมูลต่อไปนี้\n\n## ข้อมูลหลักสูตร\n{courseContent}\n\n## ข้อกำหนด\n- ทำให้ผู้อ่านเข้าใจได้เร็วว่าเป็นหลักสูตรอะไร เหมาะกับใคร และน่าสนใจอย่างไร โดยเน้นประโยชน์ที่ผู้อ่านจะได้รับ\n- คงชื่อหลักสูตรที่ถูกต้อง มีพาดหัวเด่นหนึ่งข้อความและประเด็นสำคัญ 2–3 ข้อ เลือกผลการเรียนรู้ ประสบการณ์ ผลงาน หรือจุดเด่นให้เหมาะกับหลักสูตร ไม่ต้องฝืนให้ครบ\n- ใช้ภาษาธรรมชาติและเป็นรูปธรรม ลดศัพท์เฉพาะและย่อหน้ายาว ข้อความกระชับ มีลำดับชัด อ่านง่ายบนมือถือ\n- ค่าเริ่มต้นเป็นแนวตั้ง 3:4 และสไตล์ที่เหมาะกับหลักสูตร ใช้ความต่างสีชัดเจน พื้นที่ว่างเพียงพอ และสีเน้นเล็กน้อย เพื่อให้ยังมองออกว่าเป็นเรื่องอะไรเมื่อเป็นภาพย่อ\n- จัดข้อความและการออกแบบอย่างเหมาะสม หากขาดข้อมูลที่จำเป็นต่อการตัดสินใจ ให้ถามฉัน\n- ใช้เฉพาะข้อมูลที่มีหลักฐาน ไม่แต่งข้อเท็จจริงหรือสัญญาผลลัพธ์\n- ใช้ลิงก์เดิม ใส่คิวอาร์โค้ดจริงเฉพาะเมื่อสร้างและตรวจสอบได้ มิฉะนั้นให้เว้นพื้นที่โดยไม่วาดโค้ดปลอม\n\n## ผลงานที่ต้องการ\nสร้างโปสเตอร์หนึ่งภาพ พร้อมข้อความประกอบสั้น ๆ และลิงก์หลักสูตรเดิม\nหากสร้างภาพไม่ได้ ให้ส่งข้อความโปสเตอร์และพรอมต์ที่นำไปใช้กับเครื่องมือสร้างภาพได้ทันที",
   "posterPromptLabel": "พรอมต์โปสเตอร์",
   "profileOnboardingManagement": "คำแนะนำเพื่อปรับให้เหมาะกับคุณ",
@@ -53,7 +53,7 @@
   "share": "แชร์",
   "shareContentCopied": "คัดลอกเนื้อหาสำหรับแชร์แล้ว",
   "shareCourse": "แชร์หลักสูตร",
-  "shareCourseLink": "แชร์ลิงก์หลักสูตร",
+  "shareCourseLink": "แชร์ลิงก์",
   "shareCourseMessage": "ขอแนะนำให้มาเรียน “{courseName}” และสัมผัสประสบการณ์การเรียนรู้แบบตัวต่อตัวที่ปรับให้เหมาะกับคุณ",
   "shareFailed": "แชร์ไม่สำเร็จ โปรดลองอีกครั้งภายหลัง",
   "shifu": "หลักสูตร",
