+++ src/i18n/zh-CN/common/core.json
@@ -35,8 +35,8 @@
   "packageManagement": "套餐管理",
   "pageResourcesUpdatedRefreshing": "页面资源已更新，正在自动刷新...",
   "posterCopy": "复制海报提示词",
-  "posterCopyFailed": "复制失败 请手动复制",
-  "posterNextStep": "已复制，请给到你的 AI 助理生成课程海报",
+  "posterCopyFailed": "复制失败，请手动复制",
+  "posterNextStep": "已复制，粘贴给 AI 助理生成海报",
   "posterPrompt": "请根据以下课程资料，直接生成一张适合社交媒体分享的课程海报。\n\n## 课程资料\n{courseContent}\n\n## 生成要求\n- 让读者快速看懂：这是什么课、适合谁、为什么值得了解，突出读者收益。\n- 保留准确的课程名称，提炼一个醒目的主标题和 2—3 个重点；根据课程特点选择学习收获、体验、作品或课程特色，不必凑满。\n- 用自然、具体的语言，少用术语和长段落。文字精简、层次清晰、手机上易读。\n- 默认 3:4 竖版，视觉风格与课程气质相符。使用清晰对比、足够留白和少量强调色，确保缩略图中仍能辨认主题。\n- 合理安排文案和设计；如果缺少作出决定所必需的信息，向我询问。\n- 只使用资料支持的信息，不编造事实或效果承诺。\n- 使用原课程链接。能生成并验证真实二维码时加入二维码；否则留出位置，不画假码。\n\n## 输出\n直接生成一张完整海报，不要只给设计方案。海报之外附一句简短的配图转发文案和原课程链接。\n如果不能生成图片，请给出海报文案和可直接用于生图的提示词。",
   "posterPromptLabel": "海报提示词",
   "profileOnboardingManagement": "个性化引导",
@@ -53,7 +53,7 @@
   "share": "分享",
   "shareContentCopied": "分享内容已复制",
   "shareCourse": "分享课程",
-  "shareCourseLink": "分享课程地址",
+  "shareCourseLink": "分享链接",
   "shareCourseMessage": "推荐你来学习《{courseName}》，体验一对一个性化学习。",
   "shareFailed": "分享失败，请稍后重试",
   "shifu": "课程",
