+++ src/web/src/components/course-share/index.ts
@@ -1,4 +1,5 @@
 export { CourseShareButton } from './CourseShareButton';
+export { CourseShareMenu } from './CourseShareMenu';
 export type {
   CourseShareButtonProps,
   CourseShareSurface,
