+++ src/web/src/components/header/Header.tsx
@@ -53,7 +53,7 @@ import {
   formatLessonRelativeTime,
   parseLessonHistoryDate,
 } from '@/lib/lesson-history-time';
-import { TeacherCourseShareButton } from '@/components/course-share/TeacherCourseShareButton';
+import { CourseShareMenu } from '@/components/course-share';
 import { normalizeCourseShareUrl } from '@/lib/courseShare';
 
 const publishModeIcons: Record<LearningMode, LucideIcon> = {
@@ -493,7 +493,7 @@ const Header = ({
       <div className='flex flex-row items-center'>
         <Preivew targetId={previewTargetId} />
         {currentShifu?.bid ? (
-          <TeacherCourseShareButton
+          <CourseShareMenu
             key={currentShifu.bid}
             courseTitle={currentShifu.name || ''}
             courseDescription={currentShifu.description}
